% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semaforos_get.R
\name{semaforos_get}
\alias{semaforos_get}
\title{Get Sem치foro Epidemiol칩gico Data from Mexico COVID-19 Response}
\usage{
semaforos_get(format = c("long", "wide"))
}
\arguments{
\item{format}{Either \code{"long"} (the default), or \code{"wide"} to specify the
format of the output \code{data.frame}}
}
\value{
\itemize{
\item if \code{format = "long"}, a \code{data.frame} with 4 columns:
\itemize{
\item \code{estado} which contains the name of the state to which the semaforo
in the row applies
\item \code{year} the four-digit year (\verb{\%Y} in \code{\link[=strptime]{strptime()}})
\item \code{week} the two-digit week, as specified in the ISO8601 (\verb{\%W} in
\code{\link[=strptime]{strptime()}}
\item \code{color} the color of the semaforo, indicating the alert level
}
\item if \code{format = "wide"}, a \code{data.frame} with \code{estado} as the first column
and then one column for each week, with the column names indicating the week
in the format \verb{\%Y-W\%W} (e.g., "2021-W02")
}
}
\description{
Get Sem치foro Epidemiol칩gico Data from Mexico COVID-19 Response
}
\examples{
sw <- semaforos_get(format = "wide")
head(sw)

sl <- semaforos_get(format = "long")
head(sl)

}
